requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
python-dotenv>=1.0
openai>=1.0
anthropic>=0.18
google-genai>=1.0
pyyaml>=6.0
chromadb>=0.4   # For vector RAG; use Python 3.12 or 3.13 (ChromaDB not compatible with 3.14 yet)
sentence-transformers>=2.2
